+++ e2e/packages/contracts/src/codegen/tables/StaticArray.sol
@@ -16,17 +16,22 @@ import { Schema } from "@latticexyz/store/src/Schema.sol";
 import { EncodedLengths, EncodedLengthsLib } from "@latticexyz/store/src/EncodedLengths.sol";
 import { ResourceId } from "@latticexyz/store/src/ResourceId.sol";
 
+struct StaticArrayData {
+  uint256 num;
+  uint256[3] value;
+}
+
 library StaticArray {
   // Hex below is the result of `WorldResourceIdLib.encode({ namespace: "", name: "StaticArray", typeId: RESOURCE_TABLE });`
   ResourceId constant _tableId = ResourceId.wrap(0x7462000000000000000000000000000053746174696341727261790000000000);
 
   FieldLayout constant _fieldLayout =
-    FieldLayout.wrap(0x0000000100000000000000000000000000000000000000000000000000000000);
+    FieldLayout.wrap(0x0020010120000000000000000000000000000000000000000000000000000000);
 
   // Hex-encoded key schema of ()
   Schema constant _keySchema = Schema.wrap(0x0000000000000000000000000000000000000000000000000000000000000000);
-  // Hex-encoded value schema of (uint256[])
-  Schema constant _valueSchema = Schema.wrap(0x0000000181000000000000000000000000000000000000000000000000000000);
+  // Hex-encoded value schema of (uint256, uint256[])
+  Schema constant _valueSchema = Schema.wrap(0x002001011f810000000000000000000000000000000000000000000000000000);
 
   /**
    * @notice Get the table's key field names.
@@ -41,8 +46,9 @@ library StaticArray {
    * @return fieldNames An array of strings with the names of value fields.
    */
   function getFieldNames() internal pure returns (string[] memory fieldNames) {
-    fieldNames = new string[](1);
-    fieldNames[0] = "value";
+    fieldNames = new string[](2);
+    fieldNames[0] = "num";
+    fieldNames[1] = "value";
   }
 
   /**
@@ -60,6 +66,44 @@ library StaticArray {
   }
 
   /**
+   * @notice Get num.
+   */
+  function getNum() internal view returns (uint256 num) {
+    bytes32[] memory _keyTuple = new bytes32[](0);
+
+    bytes32 _blob = StoreSwitch.getStaticField(_tableId, _keyTuple, 0, _fieldLayout);
+    return (uint256(bytes32(_blob)));
+  }
+
+  /**
+   * @notice Get num.
+   */
+  function _getNum() internal view returns (uint256 num) {
+    bytes32[] memory _keyTuple = new bytes32[](0);
+
+    bytes32 _blob = StoreCore.getStaticField(_tableId, _keyTuple, 0, _fieldLayout);
+    return (uint256(bytes32(_blob)));
+  }
+
+  /**
+   * @notice Set num.
+   */
+  function setNum(uint256 num) internal {
+    bytes32[] memory _keyTuple = new bytes32[](0);
+
+    StoreSwitch.setStaticField(_tableId, _keyTuple, 0, abi.encodePacked((num)), _fieldLayout);
+  }
+
+  /**
+   * @notice Set num.
+   */
+  function _setNum(uint256 num) internal {
+    bytes32[] memory _keyTuple = new bytes32[](0);
+
+    StoreCore.setStaticField(_tableId, _keyTuple, 0, abi.encodePacked((num)), _fieldLayout);
+  }
+
+  /**
    * @notice Get value.
    */
   function getValue() internal view returns (uint256[3] memory value) {
@@ -73,26 +117,6 @@ library StaticArray {
    * @notice Get value.
    */
   function _getValue() internal view returns (uint256[3] memory value) {
-    bytes32[] memory _keyTuple = new bytes32[](0);
-
-    bytes memory _blob = StoreCore.getDynamicField(_tableId, _keyTuple, 0);
-    return toStaticArray_uint256_3(SliceLib.getSubslice(_blob, 0, _blob.length).decodeArray_uint256());
-  }
-
-  /**
-   * @notice Get value.
-   */
-  function get() internal view returns (uint256[3] memory value) {
-    bytes32[] memory _keyTuple = new bytes32[](0);
-
-    bytes memory _blob = StoreSwitch.getDynamicField(_tableId, _keyTuple, 0);
-    return toStaticArray_uint256_3(SliceLib.getSubslice(_blob, 0, _blob.length).decodeArray_uint256());
-  }
-
-  /**
-   * @notice Get value.
-   */
-  function _get() internal view returns (uint256[3] memory value) {
     bytes32[] memory _keyTuple = new bytes32[](0);
 
     bytes memory _blob = StoreCore.getDynamicField(_tableId, _keyTuple, 0);
@@ -117,29 +141,8 @@ library StaticArray {
     StoreCore.setDynamicField(_tableId, _keyTuple, 0, EncodeArray.encode(fromStaticArray_uint256_3(value)));
   }
 
-  /**
-   * @notice Set value.
-   */
-  function set(uint256[3] memory value) internal {
-    bytes32[] memory _keyTuple = new bytes32[](0);
-
-    StoreSwitch.setDynamicField(_tableId, _keyTuple, 0, EncodeArray.encode(fromStaticArray_uint256_3(value)));
-  }
-
-  /**
-   * @notice Set value.
-   */
-  function _set(uint256[3] memory value) internal {
-    bytes32[] memory _keyTuple = new bytes32[](0);
-
-    StoreCore.setDynamicField(_tableId, _keyTuple, 0, EncodeArray.encode(fromStaticArray_uint256_3(value)));
-  }
-
   // The length of value
   uint256 constant lengthValue = 3;
-
-  // The length of value
-  uint256 constant length = 3;
 
   /**
    * @notice Get an item of value.
@@ -184,48 +187,6 @@ library StaticArray {
   }
 
   /**
-   * @notice Get an item of value.
-   * @dev Reverts with Store_IndexOutOfBounds if `_index` is out of bounds for the array.
-   */
-  function getItem(uint256 _index) internal view returns (uint256) {
-    bytes32[] memory _keyTuple = new bytes32[](0);
-
-    uint256 _byteLength = StoreSwitch.getDynamicFieldLength(_tableId, _keyTuple, 0);
-    uint256 dynamicLength = _byteLength / 32;
-    uint256 staticLength = 3;
-
-    if (_index < staticLength && _index >= dynamicLength) {
-      return (uint256(bytes32(new bytes(0))));
-    }
-
-    unchecked {
-      bytes memory _blob = StoreSwitch.getDynamicFieldSlice(_tableId, _keyTuple, 0, _index * 32, (_index + 1) * 32);
-      return (uint256(bytes32(_blob)));
-    }
-  }
-
-  /**
-   * @notice Get an item of value.
-   * @dev Reverts with Store_IndexOutOfBounds if `_index` is out of bounds for the array.
-   */
-  function _getItem(uint256 _index) internal view returns (uint256) {
-    bytes32[] memory _keyTuple = new bytes32[](0);
-
-    uint256 _byteLength = StoreCore.getDynamicFieldLength(_tableId, _keyTuple, 0);
-    uint256 dynamicLength = _byteLength / 32;
-    uint256 staticLength = 3;
-
-    if (_index < staticLength && _index >= dynamicLength) {
-      return (uint256(bytes32(new bytes(0))));
-    }
-
-    unchecked {
-      bytes memory _blob = StoreCore.getDynamicFieldSlice(_tableId, _keyTuple, 0, _index * 32, (_index + 1) * 32);
-      return (uint256(bytes32(_blob)));
-    }
-  }
-
-  /**
    * @notice Update an element of value at `_index`.
    */
   function updateValue(uint256 _index, uint256 _element) internal {
@@ -250,27 +211,125 @@ library StaticArray {
   }
 
   /**
-   * @notice Update an element of value at `_index`.
+   * @notice Get the full data.
    */
-  function update(uint256 _index, uint256 _element) internal {
+  function get() internal view returns (StaticArrayData memory _table) {
     bytes32[] memory _keyTuple = new bytes32[](0);
 
-    unchecked {
-      bytes memory _encoded = abi.encodePacked((_element));
-      StoreSwitch.spliceDynamicData(_tableId, _keyTuple, 0, uint40(_index * 32), uint40(_encoded.length), _encoded);
-    }
+    (bytes memory _staticData, EncodedLengths _encodedLengths, bytes memory _dynamicData) = StoreSwitch.getRecord(
+      _tableId,
+      _keyTuple,
+      _fieldLayout
+    );
+    return decode(_staticData, _encodedLengths, _dynamicData);
   }
 
   /**
-   * @notice Update an element of value at `_index`.
+   * @notice Get the full data.
    */
-  function _update(uint256 _index, uint256 _element) internal {
+  function _get() internal view returns (StaticArrayData memory _table) {
     bytes32[] memory _keyTuple = new bytes32[](0);
 
+    (bytes memory _staticData, EncodedLengths _encodedLengths, bytes memory _dynamicData) = StoreCore.getRecord(
+      _tableId,
+      _keyTuple,
+      _fieldLayout
+    );
+    return decode(_staticData, _encodedLengths, _dynamicData);
+  }
+
+  /**
+   * @notice Set the full data using individual values.
+   */
+  function set(uint256 num, uint256[3] memory value) internal {
+    bytes memory _staticData = encodeStatic(num);
+
+    EncodedLengths _encodedLengths = encodeLengths(value);
+    bytes memory _dynamicData = encodeDynamic(value);
+
+    bytes32[] memory _keyTuple = new bytes32[](0);
+
+    StoreSwitch.setRecord(_tableId, _keyTuple, _staticData, _encodedLengths, _dynamicData);
+  }
+
+  /**
+   * @notice Set the full data using individual values.
+   */
+  function _set(uint256 num, uint256[3] memory value) internal {
+    bytes memory _staticData = encodeStatic(num);
+
+    EncodedLengths _encodedLengths = encodeLengths(value);
+    bytes memory _dynamicData = encodeDynamic(value);
+
+    bytes32[] memory _keyTuple = new bytes32[](0);
+
+    StoreCore.setRecord(_tableId, _keyTuple, _staticData, _encodedLengths, _dynamicData, _fieldLayout);
+  }
+
+  /**
+   * @notice Set the full data using the data struct.
+   */
+  function set(StaticArrayData memory _table) internal {
+    bytes memory _staticData = encodeStatic(_table.num);
+
+    EncodedLengths _encodedLengths = encodeLengths(_table.value);
+    bytes memory _dynamicData = encodeDynamic(_table.value);
+
+    bytes32[] memory _keyTuple = new bytes32[](0);
+
+    StoreSwitch.setRecord(_tableId, _keyTuple, _staticData, _encodedLengths, _dynamicData);
+  }
+
+  /**
+   * @notice Set the full data using the data struct.
+   */
+  function _set(StaticArrayData memory _table) internal {
+    bytes memory _staticData = encodeStatic(_table.num);
+
+    EncodedLengths _encodedLengths = encodeLengths(_table.value);
+    bytes memory _dynamicData = encodeDynamic(_table.value);
+
+    bytes32[] memory _keyTuple = new bytes32[](0);
+
+    StoreCore.setRecord(_tableId, _keyTuple, _staticData, _encodedLengths, _dynamicData, _fieldLayout);
+  }
+
+  /**
+   * @notice Decode the tightly packed blob of static data using this table's field layout.
+   */
+  function decodeStatic(bytes memory _blob) internal pure returns (uint256 num) {
+    num = (uint256(Bytes.getBytes32(_blob, 0)));
+  }
+
+  /**
+   * @notice Decode the tightly packed blob of dynamic data using the encoded lengths.
+   */
+  function decodeDynamic(
+    EncodedLengths _encodedLengths,
+    bytes memory _blob
+  ) internal pure returns (uint256[3] memory value) {
+    uint256 _start;
+    uint256 _end;
     unchecked {
-      bytes memory _encoded = abi.encodePacked((_element));
-      StoreCore.spliceDynamicData(_tableId, _keyTuple, 0, uint40(_index * 32), uint40(_encoded.length), _encoded);
+      _end = _encodedLengths.atIndex(0);
     }
+    value = toStaticArray_uint256_3(SliceLib.getSubslice(_blob, _start, _end).decodeArray_uint256());
+  }
+
+  /**
+   * @notice Decode the tightly packed blobs using this table's field layout.
+   * @param _staticData Tightly packed static fields.
+   * @param _encodedLengths Encoded lengths of dynamic fields.
+   * @param _dynamicData Tightly packed dynamic fields.
+   */
+  function decode(
+    bytes memory _staticData,
+    EncodedLengths _encodedLengths,
+    bytes memory _dynamicData
+  ) internal pure returns (StaticArrayData memory _table) {
+    (_table.num) = decodeStatic(_staticData);
+
+    (_table.value) = decodeDynamic(_encodedLengths, _dynamicData);
   }
 
   /**
@@ -289,6 +348,14 @@ library StaticArray {
     bytes32[] memory _keyTuple = new bytes32[](0);
 
     StoreCore.deleteRecord(_tableId, _keyTuple, _fieldLayout);
+  }
+
+  /**
+   * @notice Tightly pack static (fixed length) data using this table's schema.
+   * @return The static data, encoded into a sequence of bytes.
+   */
+  function encodeStatic(uint256 num) internal pure returns (bytes memory) {
+    return abi.encodePacked(num);
   }
 
   /**
@@ -316,8 +383,12 @@ library StaticArray {
    * @return The lengths of the dynamic fields (packed into a single bytes32 value).
    * @return The dynamic (variable length) data, encoded into a sequence of bytes.
    */
-  function encode(uint256[3] memory value) internal pure returns (bytes memory, EncodedLengths, bytes memory) {
-    bytes memory _staticData;
+  function encode(
+    uint256 num,
+    uint256[3] memory value
+  ) internal pure returns (bytes memory, EncodedLengths, bytes memory) {
+    bytes memory _staticData = encodeStatic(num);
+
     EncodedLengths _encodedLengths = encodeLengths(value);
     bytes memory _dynamicData = encodeDynamic(value);