+++ packages/common/src/chains/mudFoundry.ts
@@ -4,7 +4,6 @@ import { MUDChain } from "./types";
 export const mudFoundry = {
   ...foundry,
   fees: {
-    // This is intentionally defined as a function as a workaround for https://github.com/wagmi-dev/viem/pull/1280
-    defaultPriorityFee: () => 0n,
+    defaultPriorityFee: 0n,
   },
 } as const satisfies MUDChain;