+++ packages/store/ts/config/v2/compat.ts
@@ -27,7 +27,7 @@ export type storeToV1<store> = store extends Store
 type schemaToV1<schema extends Schema> = { [key in keyof schema]: schema[key]["internalType"] };
 
 export type tableToV1<table extends Table> = {
-  directory: table["codegen"]["outputDirectory"];
+  directory: string;
   dataStruct: table["codegen"]["dataStruct"];
   tableIdArgument: table["codegen"]["tableIdArgument"];
   storeArgument: table["codegen"]["storeArgument"];
@@ -71,5 +71,5 @@ export function storeToV1<store>(store: conform<store, Store>): storeToV1<store>
     codegenIndexFilename: store.codegen.indexFilename,
     tables: resolvedTables,
     v2: store,
-  } as unknown as storeToV1<store>;
+  } as never;
 }