+++ packages/store/ts/config/v2/generics.ts
@@ -3,10 +3,20 @@ import { merge } from "@arktype/util";
 export type get<input, key> = key extends keyof input ? input[key] : undefined;
 
 export function get<input, key extends PropertyKey>(input: input, key: key): get<input, key> {
-  return (typeof input === "object" && input != null && hasOwnKey(input, key) ? input[key] : undefined) as get<
-    input,
-    key
-  >;
+  return (typeof input === "object" && input != null && hasOwnKey(input, key) ? input[key] : undefined) as never;
+}
+
+export type getPath<input, path extends readonly PropertyKey[]> = path extends readonly [
+  infer head,
+  ...infer tail extends PropertyKey[],
+]
+  ? head extends keyof input
+    ? getPath<input[head], tail>
+    : undefined
+  : input;
+
+export function getPath<input, path extends readonly PropertyKey[]>(input: input, path: path): getPath<input, path> {
+  return path.length ? (getPath(get(input, path[0]), path.slice(1)) as never) : (input as never);
 }
 
 export function hasOwnKey<obj, const key extends PropertyKey>(
@@ -39,5 +49,7 @@ export function mergeIfUndefined<base extends object, merged extends object>(
   const allKeys = [...new Set([...Object.keys(base), ...Object.keys(merged)])];
   return Object.fromEntries(
     allKeys.map((key) => [key, base[key as keyof base] ?? merged[key as keyof merged]]),
-  ) as mergeIfUndefined<base, merged>;
+  ) as never;
 }
+
+export type parseNumber<T> = T extends `${infer N extends number}` ? N : never;