+++ packages/store/ts/config/v2/input.ts
@@ -1,6 +1,11 @@
 import { Hex } from "viem";
-import { Codegen, Enums, TableCodegen, TableDeploy, UserTypes } from "./output";
+import { Codegen, TableCodegen, TableDeploy, UserTypes } from "./output";
 import { Scope } from "./scope";
+import { evaluate } from "@arktype/util";
+
+export type EnumsInput = {
+  readonly [enumName: string]: readonly [string, ...string[]];
+};
 
 export type SchemaInput = {
   readonly [key: string]: string;
@@ -10,6 +15,9 @@ export type ScopedSchemaInput<scope extends Scope> = {
   readonly [key: string]: keyof scope["types"];
 };
 
+export type TableCodegenInput = Partial<TableCodegen>;
+export type TableDeployInput = Partial<TableDeploy>;
+
 export type TableInput = {
   readonly schema: SchemaInput;
   readonly key: readonly string[];
@@ -17,20 +25,29 @@ export type TableInput = {
   readonly name: string;
   readonly namespace?: string;
   readonly type?: "table" | "offchainTable";
-  readonly codegen?: Partial<TableCodegen>;
-  readonly deploy?: Partial<TableDeploy>;
+  readonly codegen?: TableCodegenInput;
+  readonly deploy?: TableDeployInput;
 };
 
 export type TablesInput = {
   readonly [key: string]: Omit<TableInput, "namespace" | "name">;
 };
 
+export type CodegenInput = Partial<Codegen>;
+
 export type StoreInput = {
+  /**
+   * Directory of Solidity source relative to the MUD config.
+   * This is used to resolve other paths in the config, like codegen and user types.
+   *
+   * Defaults to `src` to match `foundry.toml`'s default. If you change this from the default, you may also need to configure foundry with the same source directory.
+   */
+  readonly sourceDirectory?: string;
   readonly namespace?: string;
   readonly tables?: TablesInput;
   readonly userTypes?: UserTypes;
-  readonly enums?: Enums;
-  readonly codegen?: Partial<Codegen>;
+  readonly enums?: EnumsInput;
+  readonly codegen?: CodegenInput;
 };
 
 /******** Variations with shorthands ********/
@@ -41,4 +58,4 @@ export type TablesWithShorthandsInput = {
   readonly [key: string]: TableInput | TableShorthandInput;
 };
 
-export type StoreWithShorthandsInput = Omit<StoreInput, "tables"> & { tables: TablesWithShorthandsInput };
+export type StoreWithShorthandsInput = evaluate<Omit<StoreInput, "tables"> & { tables: TablesWithShorthandsInput }>;