+++ templates/vanilla/packages/client/src/mud/supportedChains.ts
@@ -10,10 +10,10 @@
  *
  */
 
-import { MUDChain, latticeTestnet, mudFoundry } from "@latticexyz/common/chains";
+import { MUDChain, mudFoundry, redstone, garnet } from "@latticexyz/common/chains";
 
 /*
  * See https://mud.dev/tutorials/minimal/deploy#run-the-user-interface
  * for instructions on how to add networks.
  */
-export const supportedChains: MUDChain[] = [mudFoundry, latticeTestnet];
+export const supportedChains: MUDChain[] = [mudFoundry, redstone, garnet];